Generate basic player model
Using bundle exec hanami general model player.

We specify that we will be accepting json in our controller as that is
what we want to use. We also ignore templates in the view because we
don't need to visually present anything.

Sorry for the confusion with the spec here, it was from a tutorial and
it didn't work, but I get the idea and I will get to writing some specs
later.