Follow Hanami examples for DI
It has been argued that the previous version of DI in the diff somewhat
defeats the purpose of DI. The code would work without being passed any
dependencies as we set it anyways.

Now we follow the examples that Hanami uses for DI and set it in the
parameters. This will still allow us to influence the dependency while
testing.