Add find_by_token method to Player Repository
This allows us to find a Player entity in a repository by searching for
their token. The reason we want to do this is to allow us to use
Hanami's DSL for deleting from a repository, which only deletes by
entity ID. By finding a Player entity by their token we also find their
ID and can use it to delete them from the repository.