Add action for delete player endpoint
we quit early if we are given a nil player. An argument can be made that
by introducing nil at a lower level in the repository's find_by_token
method that we have now had to add two extra nil checks but I don't
believe this to be a huge issue at the moment. I am open to other ideas.