Add Iodine to the Gemfile
We have decided to use Iodine as the server for Hanami. The reason for
this being that Hanami does not support websockets natively. Enable to
utilize websockets we need to turn to a third party HTTP / websocket
server that has native Pub/Sub support. With Iodine we can do
everything we could previously do with Hanami but also utilize
websockets.

Here are the two resources being used:
https://github.com/boazsegev/iodine
https://github.com/boazsegev/iodine/blob/master/SPEC-WebSocket-Draft.md