Author Websocket middleware
We introduce middleware in this file to check incoming HTTP requests for
attempted websocket communications. We check this way because Iodine
publishes Websocket support using an env Hash.

We don't need to specify headers and a body to complete the websocket
handshake as Iodine takes care of it for us, hence we simply return a
status 200.

A future implementation will be an attempt to move this middleware out
of config.ru and place it elsewhere while utilizing Hanami's API to
insert the middleware into the application's middleware chain.