Generate Controller upon app start
My logic was a bit off in the previous implementation. If we generate a
Controller object when our app runs we can then work out of a single
location and generate a list of Clients for every incoming connection.

This makes more sense because each incoming connection becomes a Client
which we can build out further in the future. The Controller acts as
a hub for handling connections and messages, which we can further break
down in the future.