Create Client interactor
I wasn't sure what to name it, please make a suggestion if you can.

We don't want to clog up our Controller class with unnecessary logic.
By moving all of the logic out of the Controller class into an
interactor we allow the Controller to tell an interactor what to do
rather than doing it itself.