Re-factor BuildPayload and UpdateAllClients
We are going to slowly move into a world where we will need a factory
that generates JSON payloads based on the information we need modeled
at the time. This is a small step in that direction, but will likely
need adjustment as we need more things where we can fully build out a
factory.