Update the race status to all Clients upon update
When we receive a message we need to authenticate that the incoming
data is updating the correct Clients state because we want to update the
correct Client as well as not allow other Clients to change each others
attributes.

When we receive a position update from a player we want to update all
Clients with the new race state to keep the rendered board up to date.