Store path info from connections HTTP request
When a connection is established to the server we receive a hash that
details the HTTP Request. We can access this and save the path info to
use as a server 'room'. This will allow us to generate a URL on the
React client and associate connected clients with that room URL.

example: www.playgame.com/34

where 34 is the lobby ID.