Generate FetchAllRooms action
We want to be able to visit an endpoint to see all the Room entities
that have been generated. We will inevitably want to hit a POST endpoint
to generate a Room entity, but this index will provide us with a way to
see all Room entities that have been generated.