Update Controller to setup and update rooms
When a new connection comes in we want to add it to the room that it
belongs to per the PATH INFO we receive in the request. We want to
create a new room if the room being requested doesn't exist so that we
can add all future connections for that room, to that room.