Add room and position columns to Player entity
Due to concurrency we can't maintain in-memory state across all threads.
We will attempt to keep track of Player states via the database so that
we can build race payloads serverside to keep work off of the front end.

We need to figure out a way to do this because the front end will be
drastically slow on slower computers if we are maintaining state there.