Create PlayersRooms Repository and query methods
Based on the migration for the players_rooms table, we need to be able
to query for all players in a given room or a single given player in a
given room.

We need to be able to do this so that we can update individual players
or send updates to all players in a room.