Update RacePayload
We no longer have the concept of a Client object. Therefore we need to
be able to build a RacePayload object based off of information in the
database given a room.

This implementation allows us to build the same payload that we were
using previously with the concept of Client and Room objects. This
ensures that we don't need to change our front-end code.