Update RaceUpdate
We no longer have the concept of Client and Room objects. Therefore we
need to be able to send race updates to all connections attached to a
given room. This is achievable with Iodine's pub/sub methods on our
connection object because it stores all connections to a room for us,
allowing us to not have to maintain connections in-memory.