Implement HandleNewConnection service object
We no longer have the concept of Client and Room objects. By using
Iodine's pub/sub implementation we can avoid concurrency issues and
subscribe new connections to the room provided in the env hash from the
request.

Because Room entity's are created before a Player entity is generated
for a given connection, we find the room that was created and build an
associaton between the new Player entity and the Room. This allows us
to perform operations like race updates to all players in a room or
add new players to a room etc.

When a new connection comes in we need to let it know what its ID is
for all future updates it is to send and receive. Once this is done we
also need to send the latest race update to all players in the room
that the new connection has joined so that all clients can render the
most up to date race.