Implement HandleMessage service object
We no longer have the concept of Client and Room objects. By using
Iodine's pub/sub implementation we can avoid concurrency issues and
send updates to specific rooms that connections have subscribed to
upon connecting.

At the moment we only have two kinds of messages to worry about. When
we receive a race update we need to alert every player in the associated
room that a players position has changed so that they can render the
most up to date race. Because we have set up a players_rooms association
table in the database we can update a given player's position there.

When we receive a timer update we also need to alert every player in
the associated room that the game has been started so that they can
also start their race timers to stay in sync.