Update controller
We no longer have the concept of Client and Room objects. By using
Iodine's pub/sub implementation we can avoid concurrency issues and
subscribe new connections to the room provided in the env hash from the
request.

Because of this we can also send updates to specific rooms that
connections have subscribed to upon connecting.