Add .env.development to .gitignore
We will be implementing OAuth with Slack that will involve storing
secret information in the development ENV file. To avoid publicizing
this information on GitHub we won't allow it to be uploaded.

Moving forward we will have to reset all keys currently stored in
.env.development when we move to production as they've been in this
repo for a while and that's probably not a good thing.