Add FetchOauth interactor to auth with slack
We need to persist Player entities. To do this we will OAuth with slack
and build Players with the information. This also provides us with
security, as we can ensure that information is accurate and Players are
who they say they are.

I will have to re-visit this to write a spec once I learn how the VCR
gem works.