Add controller for Slack OAuth endpoint
We have an endpoint for our client to hit, giving us information to
use to make requests to Slack oauth. This keeps secret information off
of the client where it could be viewed maliciously. This also allows us
to return authenticated information to our client to use throughout
its session.

I will have to re-visit this to write a spec when I know how the VCR
gem works.