Add View for Slack OAuth endpoint
We need to be able to authenticate all game updates and requests. To do
this we will give our client an expiring UUID and let it know whether
its been authenticated or not.

I will have to re-visit this to write a spec when I understand how the
VCR gem works.