Implement UpdatePlayer interactor
When a person goes to play the game they will require a new UUID upon
being authenticated if they want to be able to perform game updates.

We will implement a way for UUIDs to expire, but for now every time a
user is authenticated they will be given a new one.