Add VCR Ruby gem dependency
We need a way to mock our requests to the Slack API. VCR supports
Faraday, the gem that we are using to make these requests.

With this gem we will be able to mock responses from the Slack API once
and not make any subsequent requests. We will also be able to use VCR to
redact sensitive information from our specs that we need for testing but
don't want to be present in the github repository.