Implement RaceState Model
We can use a Model to represent the previously known payloads. This
will clean up the rest of our code in how we go about distinguishing
between payload types, both on the server and client side.

Further implementation might split the JSON serialization into a
separate service as the type to send is an API related concern that the
model doesn't need to know about.