Implement StatsState Model
We can represent the stats state that we want to send to the client with
this model. The model holds a hash of players that contains their stats.
This allows us to continue using the pattern that the race position
update uses. This also makes implementation on the client side much
easier because of the similarity.