Update UpdatePlayerRoom interactor
Now that we aren't only updating players' positions but also their
stats in a room, we need to update our player_room entities in different
ways.

This allows us to send this interactor different things to achieve
different results, but we might want to separate this into two different
interactors to be more explicit and separate concerns.