Implement HandleStatsUpdate Handler
By separating this service into a Handler we can separate some concern
from the update service that sends the players stats to the clients.

The only thing we are concerned with here is updating the players stats
in the database and routing accordingly for the state update to be made
for the clients.