Add columns to rooms table
We need to stop players from joining games that are in progress. We
already have ways of knowing when a game has been started, so we can
just latch onto those to use this column and deny and late comers to a
game.

We can also use the updated_at column to expire rooms as well so that
you can't go back and play old games to get better stats.