Add interactor to calculate players stats
Given player_room records we want to determine a number of variables.
While not perfect, this gets the job done. The calculate_stats method
could likely be written nicer with some more thinking, but for now I
want to get some results on the client and this sets up the payload I
would like to receive on the client.