Add controller for player_rooms#show
We need an endpoint that our client can hit to retrieve player stats.

This endpoint requires a UUID for information to be retrieved. This is
important right now, but moving forward we will want any client to
request information for any other player in their workspace. This will
allow a leaderboard to be displayed, or all player stats for each game
played etc. The sky is the limit.