Add rounding to data sent by endpoint
There was a descrepency in the client where data received could look
like "96.6666666667" for things like accuracy. Because we apply this
rounding elsewhere on the server, we can opt to do it here as well,
rather than on the client side. This way we know where to look if things
go wrong.