Reverse player_room records for building stats
The client presents the game history in a scrollable box. It makes sense
to me to display the latest game first. By reversing the array of
game records we achieve this as the client renders the games from
position 0.

I opted to do the reverse here instead of in the repository because
there is a chance we might want to use the repository method without
reversing the games played at some point.